Delimit if/else branches with braces instead of \else
When `cleveref 0.21.4` is loaded together with `fmtcount` 3.05 with the french language, e.g with
```
\documentclass[12pt,french]{article}
\usepackage{babel}
\usepackage{cleveref}
\usepackage{fmtcount}
\begin{document}
\ordinalstringnum{1}
\end{document}
```
the result is
```
Incomplete \ifnum; all text was ignored after line ...
```
This is probably due to a command like \@tempa or similar that cleveref redefines to \iftrue or \iffalse somewhere for its usage, which leads the fast conditional skipping of TeX astray. Use braced arguments and \@firstoftwo/\@secondoftwo instead, which fixes the problem.