Fix spurious space in Arabic \@ord@unitstringarabic.
* trunk/arabic-implementation.tex: New file for arabic
documentation. Here we need UTF-8 encoding, and xetex is a better
choice for compilation engine, that is why a separate file is
preferable.

* trunk/fc-arabic.def: Place all DocTeX documented macrocode
environement between \ifxetex…\fi, so that it is compiled only
with XeTeX. Start to translate code documentation to French, as it
is the common language used with contributors to Arabic
module. Suppress spurious spaces in \@ord@unitstringarabic.

* trunk/test/sample-arabic.tex: Add numbers up to 1019 by step of
7.